Add files via upload
This process took me a while to get the results I wanted. The hardest part was trying to get compared number to print versus the compared number which in this case would be -1. I added notes within my coded descripting why i used the methods the way i did. This was a challenging assignment however i feel I learned a lot in the process.